src/main/java/com/springbook/application/infrastructure/security/ApplicationUserDetailsService.java
package com.springbook.application.infrastructure.security;

import com.springbook.application.user.User;
import com.springbook.application.user.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class ApplicationUserDetailsService implements UserDetailsService {

    private final UserRepository userRepository;

    @Autowired
    public ApplicationUserDetailsService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userRepository.findByEmailIgnoreCase(username)
                                  .orElseThrow(() -> new UsernameNotFoundException(String.format("User with email %s could not be found", username)));

        return new ApplicationUserDetails(user);
    }
}
